Refactor: Remove ride path drawing from MapPathScreen
This commit removes the code responsible for drawing the ride path on the `MapPathScreen`. The ride path was previously drawn as a series of red lines connecting GPS fixes.

The section of code that iterated through `fixes` to create path segments and draw lines using `drawLine` has been deleted. Cafe markers will still be drawn.