Refactor: Use `app_name` for AshBike Top Bar Title
This commit updates the `TopBarForCurrentRoute` composable in the AshBike application to use the generic `R.string.app_name` resource for the title of the home screen, instead of the more specific `R.string.top_bar_title_ashbike`.

Key changes:

-   **`applications/ashbike/src/main/java/com/ylabz/basepro/applications/bike/ui/navigation/main/TopBarForCurrentRoute.kt`**:
    -   Changed the `titleResId` for `BikeScreen.HomeBikeScreen.route` from `R.string.top_bar_title_ashbike` to `R.string.app_name`.

This change promotes the use of a common string resource for the application's name, potentially simplifying branding or renaming efforts in the future.