Build: Update minSdk to use version catalog
This commit updates the `minSdk` value in the `build.gradle.kts` files for various application and feature modules.
The hardcoded `minSdk = 32` has been replaced with `minSdk = libs.versions.minSdk.get().toInt()` to utilize the version defined in the Gradle version catalog.

This change centralizes the `minSdk` definition, making it easier to manage and update across the project.