Refactor: Pass `isExpandedScreen` down to NavGraph
This commit refactors the navigation structure to pass the `isExpandedScreen` boolean from `MainScreen` down into the `PhotoDoNavGraph`.

Previously, `isExpandedScreen` was hardcoded to `true` within the `entryProvider` of the `PhotoDoNavGraph`. This change removes the hardcoded value and allows the `PhotoDoNavGraph` to be aware of the actual screen size, enabling more responsive and adaptive layouts.